requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Babel==2.9.1
bleak==0.12.1
certifi==2018.8.24
chardet==3.0.4
crcmod==1.7
cryptography==2.6.1
dbus-next==0.2.3
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
gateway==1.2.0
idna==2.6
imagesize==1.3.0
importlib-metadata==4.10.1
iniconfig==1.1.1
Jinja2==3.0.3
keyring==17.1.1
keyrings.alt==3.1.1
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.3
paho-mqtt==1.6.1
pipreqs==0.4.11
pluggy==1.0.0
psutil==5.5.1
ptyprocess==0.7.0
py==1.11.0
pycrypto==2.6.1
Pygments==2.11.2
PyGObject==3.30.4
pyparsing==3.0.7
pytest==6.2.5
python-apt==1.8.4.3
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
regex==2022.1.18
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.7
toml==0.10.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.24.1
yarg==0.1.9
zipp==3.7.0
